requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
pyyaml